Implemented everything except for "Analiza nationala"
Debating whether it would be more worth to rewrite everything from
scratch in such a way that access to regions is more efficient.
Or if i should just implement it as-is and get a working product
Still not very sure which one of them will be easier.
Rewriting the whole project in a much more "OOP way" would be a good
exercise and also an opportunity to understand OOP much better

Signed-off-by: Panait Teodor Dimitrie <teodor.panait@stud.acs.upb.ro>